requirements.txt
aiofiles==23.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
botocore==1.33.13
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
diff-match-patch==20230430
distlib==0.3.8
dj-database-url==2.2.0
Django==4.2.3
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-heroku==0.3.1
django-import-export==4.1.1
django-jazzmin==3.0.0
django-modeltranslation==0.18.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
entrypoints==0.4
ezdxf==1.4.2
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.14.0
fonttools==4.39.3
frozenlist==1.4.1
fsspec==2023.4.0
gradio==3.27.0
gradio_client==0.1.3
gunicorn==23.0.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
idna==3.7
inflection==0.5.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
linkify-it-py==2.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.5
numpy==1.24.2
opencv-python==4.10.0.82
orjson==3.8.10
packaging==24.1
pandas==2.0.0
pillow==10.4.0
pipenv==2023.12.1
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==1.10.7
pydot==1.4.2
pydotplus==2.0.2
pydub==0.25.1
pygraphviz==1.11
PyJWT==2.9.0
pyotp==2.9.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.4.3
requests==2.32.3
rpds-py==0.20.0
seaborn==0.12.2
semantic-version==2.10.0
shapely==2.1.1
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.0
sqlparse==0.5.1
starlette==0.26.1
tablib==3.5.0
toolz==0.12.0
tqdm==4.65.0
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.21.1
virtualenv==20.26.2
virtualenv-clone==0.5.7
websockets==11.0.2
whitenoise==6.7.0
yarl==1.9.4
